Implement: login/signup users

+++ routes/movie.js
@@ -1,14 +1,17 @@
 const router = require("express").Router();
 const Movie = require("../models/Movie");
+const verifyToken = require("../middlewares/verifyToken");
 
-router.get("/movie/:id?", async (req, res) => {
+router.get("/movie/:id?", verifyToken, async (req, res) => {
   const movieId = req.params.id;
   try {
     if (movieId) {
-      const movie = await Movie.findById(movieId);
+      const movie = await Movie.findOne({ _id: movieId, user_id: req.user.id });
       res.json(movie);
     } else {
-      const movie = await Movie.find().sort({ updatedAt: -1 });
+      const movie = await Movie.find({ user_id: req.user.id }).sort({
+        updatedAt: -1,
+      });
       res.json(movie);
     }
   } catch (error) {
@@ -17,14 +20,17 @@ router.get("/movie/:id?", async (req, res) => {
   }
 });
 
-router.post("/movie", async (req, res) => {
+router.post("/movie", verifyToken, async (req, res) => {
   try {
-    const exists = await Movie.findById(req.body?._id);
+    const exists = await Movie.findOne({
+      show_id: req.body?.show_id,
+      user_id: req.user.id,
+    });
 
     if (exists) {
       res.sendStatus(409);
     } else {
-      const movie = await Movie.create(req.body);
+      const movie = await Movie.create({ ...req.body, user_id: req.user.id });
 
       res.status(201).json(movie);
     }
@@ -34,14 +40,14 @@ router.post("/movie", async (req, res) => {
   }
 });
 
-router.put("/movie/:id", async (req, res) => {
+router.put("/movie/:id", verifyToken, async (req, res) => {
   const movieId = req.params.id;
   const { body } = req;
 
   if (!movieId || !body) res.sendStatus(204);
 
   try {
-    const movie = await Movie.findOne({ _id: movieId });
+    const movie = await Movie.findOne({ _id: movieId, user_id: req.user.id });
 
     if (!movie) return res.sendStatus(204);
 
@@ -55,12 +61,16 @@ router.put("/movie/:id", async (req, res) => {
   }
 });
 
-router.delete("/movie/:id", async (req, res) => {
+router.delete("/movie/:id", verifyToken, async (req, res) => {
   const movieId = req.params.id;
 
   if (!movieId) res.sendStatus(400);
 
   try {
+    const movie = await Movie.findById(movieId);
+
+    if (movie.user_id !== req.user.id) return res.sendStatus(403);
+
     await Movie.deleteOne({ _id: movieId });
     res.sendStatus(200);
   } catch (error) {
