+++ routes/show.js
@@ -1,14 +1,18 @@
 const router = require("express").Router();
 const Show = require("../models/Show");
+const verifyToken = require("../middlewares/verifyToken");
 
-router.get("/show/:id?", async (req, res) => {
+router.get("/show/:id?", verifyToken, async (req, res) => {
   const showId = req.params.id;
   try {
     if (showId) {
-      const show = await Show.findById(showId);
+      const show = await Show.findOne({ _id: showId, user_id: req.user.id });
+
       res.json(show);
     } else {
-      const shows = await Show.find().sort({ updatedAt: -1 });
+      const shows = await Show.find({ user_id: req.user.id }).sort({
+        updatedAt: -1,
+      });
       res.json(shows);
     }
   } catch (error) {
@@ -17,11 +21,11 @@ router.get("/show/:id?", async (req, res) => {
   }
 });
 
-router.get("/show/:id/complete", async (req, res) => {
+router.get("/show/:id/complete", verifyToken, async (req, res) => {
   const showId = req.params.id;
 
   try {
-    const show = await Show.findOne({ _id: showId });
+    const show = await Show.findOne({ _id: showId, user_id: req.user.id });
 
     if (!show) return res.sendStatus(204);
 
@@ -36,14 +40,17 @@ router.get("/show/:id/complete", async (req, res) => {
   }
 });
 
-router.post("/show", async (req, res) => {
+router.post("/show", verifyToken, async (req, res) => {
   try {
-    const exists = await Show.findById(req.body?._id);
+    const exists = await Show.findOne({
+      show_id: req.body?.show_id,
+      user_id: req.user.id,
+    });
 
     if (exists) {
       res.sendStatus(409);
     } else {
-      const show = await Show.create(req.body);
+      const show = await Show.create({ ...req.body, user_id: req.user.id });
 
       res.status(201).json(show);
     }
@@ -53,7 +60,7 @@ router.post("/show", async (req, res) => {
   }
 });
 
-router.put("/show/:id", async (req, res) => {
+router.put("/show/:id", verifyToken, async (req, res) => {
   const showId = req.params.id;
   const { body } = req;
 
@@ -63,6 +70,8 @@ router.put("/show/:id", async (req, res) => {
     const show = await Show.findOne({ _id: showId });
 
     if (!show) return res.sendStatus(204);
+
+    if (show.user_id !== req.user.id) return res.sendStatus(403);
 
     if (Number.isInteger(body.seasonsWatched))
       show.seasonsWatched = body.seasonsWatched;
@@ -80,12 +89,16 @@ router.put("/show/:id", async (req, res) => {
   }
 });
 
-router.delete("/show/:id", async (req, res) => {
+router.delete("/show/:id", verifyToken, async (req, res) => {
   const showId = req.params.id;
 
   if (!showId) res.sendStatus(400);
 
   try {
+    const show = await Show.findById(showId);
+
+    if (show.user_id !== req.user.id) return res.sendStatus(403);
+
     await Show.deleteOne({ _id: showId });
     res.sendStatus(200);
   } catch (error) {
